new file with process d instead of function in manager
